test(credentials,schema): move pure logic onto the coverage gate
The *_integration.go / *_interactive.go suffix exempts a whole file from
the 80% coverage gate (SonarCloud + pre-commit) by filename. Pure,
already-tested logic was hiding behind that suffix alongside live cloud
constructors. Split the pure part into gated .go siblings so it counts.

- aws_integration.go -> aws_signing.go (95.8%); NewAWSCredential* stay live
- azure_integration.go -> azure.go (95.7%); NewAzureCredential* stay live
- gcp_integration.go -> gcp.go (96.0%); NewGCPCredential* stay live
- resolver_platform_integration.go -> resolver_platform.go (100%)
- prompt/schema/schema_interactive.go -> prompt/schema/schema.go + schema_test.go (100%)

Moving getCanonicalHeaders onto the gate surfaced two latent staticcheck
issues (WriteString(fmt.Sprintf(...))); converted to fmt.Fprintf.

All existing sibling tests still pass; -race green.
