+++ runtime/credentials/aws_signing_test.go
