+++ runtime/credentials/resolver_platform.go
